# src/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:8077",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {},
  "consumes": [
    "application/json"
  ],
  "produces": [],
  "paths": {
...
    "/recipes/migrate_all": {
      "get": {
        "tags": [
          "Recipe"
        ],
        "description": "Endpoint to transport the images of all the remaining recipes to S3 storage",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "All migrated"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "Permission Denied!"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "406": {
            "description": "Not Acceptable",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "Database error: [error]"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/recipes/migrate_to_s3": {
      "get": {
        "tags": [
          "Recipe"
        ],
        "description": "Endpoint to transport the image of a recipe to S3 storage",
        "parameters": [
          {
            "name": "recipe_name",
            "in": "body",
            "description": "Recipe name",
            "required": true,
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "recipe_name": {
                  "type": "string",
                  "example": "Menemen"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "All migrated"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "Permission Denied!"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "Recipe was not found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "406": {
            "description": "Not Acceptable",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "Database error: [error]"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/recipes/nutrition_filter": {
      "post": {
        "tags": [
          "Recipe"
        ],
        "description": "Endpoint to find a recipe with nutrition constraints",
        "parameters": [
          {
            "name": "type",
            "in": "body",
            "description": "Recipe name",
            "required": true,
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "Breakfast"
                }
              }
            }
          },
          {
            "name": "calories",
            "in": "body",
            "description": "Recipe name",
            "required": true,
            "type": "integer",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "number",
                  "example": 2000
                }
              }
            }
          },
          {
            "name": "protein",
            "in": "body",
            "description": "Recipe name",
            "required": true,
            "type": "integer",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "number",
                  "example": 200
                }
              }
            }
          },
          {
            "name": "carbs",
            "in": "body",
            "description": "Recipe name",
            "required": true,
            "type": "integer",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "number",
                  "example": 100
                }
              }
            }
          },
          {
            "name": "fat",
            "in": "body",
            "description": "Recipe name",
            "required": true,
            "type": "integer",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "number",
                  "example": 50
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "imgPath": {
                  "type": "string",
                  "example": ""
                },
                "servingSize": {
                  "type": "number",
                  "example": 4
                },
                "instructions": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "likedUsers": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "oid": {
                        "type": "string",
                        "example": "user_id"
                      }
                    },
                    "required": [
                      "oid"
                    ]
                  }
                },
                "numberOfLikes": {
                  "type": "number",
                  "example": 2
                },
                "type": {
                  "type": "string",
                  "example": "lunch"
                },
                "name": {
                  "type": "string",
                  "example": "C"
                },
                "prepTime": {
                  "type": "number",
                  "example": 55
                },
                "cookTime": {
                  "type": "number",
                  "example": 35
                },
                "ingredients": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "object",
                        "properties": {
                          "oid": {
                            "type": "string",
                            "example": "ingredient_id2"
                          }
                        },
                        "required": [
                          "oid"
                        ]
                      },
                      "text": {
                        "type": "string",
                        "example": "1 tbsp of sugar and a lot of salt"
                      },
                      "unit": {
                        "type": "string",
                        "example": "tbsp"
                      },
                      "name": {
                        "type": "string",
                        "example": "sugar"
                      },
                      "originID": {
                        "type": "object",
                        "properties": {
                          "oid": {
                            "type": "string",
                            "example": "ingredient_original_id2"
                          }
                        },
                        "required": [
                          "oid"
                        ]
                      }
                    }
                  }
                },
                "__v": {
                  "type": "number",
                  "example": 1
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "Permission Denied!"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "Recipe with required nutritions was not found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "406": {
            "description": "Not Acceptable",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "Something went wrong with find a recipe"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/recipes": {
      "post": {
        "tags": [
          "Recipe"
        ],
        "description": "Endpoint to add a recipe",
        "parameters": [
          {
            "name": "name",
            "in": "body",
            "description": "Recipe name",
            "required": true,
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Menemen"
                }
              }
            }
          },
          {
            "name": "prepTime",
            "in": "body",
            "description": "Recipe preperation time",
            "required": true,
            "type": "integer",
            "schema": {
              "type": "object",
              "properties": {
                "prepTime": {
                  "type": "number",
                  "example": 30
                }
              }
            }
          },
          {
            "name": "cookTime",
            "in": "body",
            "description": "Recipe cook time",
            "required": true,
            "type": "integer",
            "schema": {
              "type": "object",
              "properties": {
                "cookTime": {
                  "type": "number",
                  "example": 50
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "Recipe is added"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "Did not add the recipe"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "tags": [
          "Recipe"
        ],
        "description": "Endpoint for a user to get a Recipe / meal recommendation. If user has saved information about the preferred number of meals per day, that \\\rnumber will be used. Otherwise if limit is passed as a path parameter it will be used. If that is also not passed, then a constant number 3 will be used.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "the List of all recipe names is returned"
          },
          "400": {
            "description": "Either no recipe or no authorization"
          }
        }
      }
    },
    "/recipes/recommendation": {
      "get": {
        "tags": [
          "Recipe"
        ],
        "description": "Endpoint for a user to get a Recipe / meal recommendation. If user has saved information about the preferred number of meals per day, that \\\rnumber will be used. Otherwise if limit is passed as a path parameter it will be used. If that is also not passed, then a constant number 3 will be used.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "maximum number of recipes to return",
            "required": false,
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "limit": {
                  "type": "string",
                  "example": "2"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of recipe names is returned"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/recipes/prefill": {
      "post": {
        "tags": [
          "Recipe"
        ],
        "description": "Endpoint to add the recipes in local to db",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "Recipes are added!"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
...
}